Consolidate SSH/SLURM helper duplication across pipeline/*_remote.py
_env_int, _env_text, _config_text, _resolve_ssh_options, _assert_ssh_reachable,
and REMOTE_FAILURE_PREFIXES were byte-identical (modulo cosmetic whitespace/
naming, verified with a diff pass) copy-pasted across all four remote-execution
modules (interproscan_remote.py, colabfold_remote.py, ligq_remote.py,
structures_remote.py) and slurm_remote_command.py itself. Promote them into
slurm_remote_command.py -- already the shared module, previously used only by
run_pipeline_direct.py's direct SLURM path -- and have all four import from
there instead of reimplementing.

_open_remote_session/_close_remote_session (SSH+SCP+SFTP session open, and
close-everything-defensively) and _exec_remote (the exec-command-and-decode
wrapper) were also functionally identical across ligq_remote.py,
structures_remote.py, and colabfold_remote.py -- promote those too.
interproscan_remote.py manages its session differently (opens ssh/scp/sftp
inline, closes each independently in its own try/except) and wasn't touched
for this piece, since forcing it into the same shape isn't a mechanical move.

Explicitly NOT touched: each file's submit->poll->classify-failure state
machine. A structured investigation (see the audit this commit is based on)
confirmed these are genuinely different behaviorally, not just duplicated --
completion detection differs (scp-probe vs. DONE-marker vs. sacct state),
only interproscan/colabfold do squeue pending-reason lookups + completion
grace periods + friendly resource-constraint messages, and colabfold's SSH
session lifecycle is fundamentally different (opened/closed per GPU/CPU
retry attempt, not once per run, to support its parallel ThreadPoolExecutor
fallback orchestration). Forcing one shared poll-loop implementation across
all four would need so many behavioral hooks that it stops being meaningful
deduplication, and the risk of silently altering one stage type's real
production behavior isn't worth it without live verification per stage --
this matches the same judgment already applied to the .pocket-card CSS
component in the earlier CSS consolidation pass.

Verified by ast.parse + ruff (paramiko/scp aren't installed in this
environment to actually import-test); every stage type needs to be
triggered on the cluster to confirm live behavior, since none of this is
testable without real SSH/SLURM access.
