+++ pipeline/colabfold_remote.py
@@ -17,89 +17,22 @@ import gzip
 import os
 import shlex
 import shutil
-import socket
 import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from dataclasses import dataclass
 from datetime import datetime
 
-import paramiko
-from scp import SCPClient
-
-from tpweb.services.slurm_messages import classify_slurm_resource_message
-
-
-REMOTE_FAILURE_PREFIXES = (
-    "FAILED",
-    "CANCELLED",
-    "TIMEOUT",
-    "OUT_OF_MEMORY",
-    "NODE_FAIL",
+from slurm_remote_command import (
+    REMOTE_FAILURE_PREFIXES,
+    _assert_ssh_reachable,
+    _close_remote_session,
+    _config_text,
+    _env_int,
+    _env_text,
+    _open_remote_session,
+    _resolve_ssh_options,
 )
-
-
-def _env_int(name, default):
-    raw = os.getenv(name)
-    if raw is None:
-        return default
-    try:
-        return int(raw.strip())
-    except (TypeError, ValueError):
-        return default
-
-
-def _env_text(name, default=None):
-    raw = os.getenv(name)
-    if raw is None:
-        return default
-    text = str(raw).strip()
-    return text or default
-
-
-def _assert_ssh_reachable(host, port, timeout_seconds):
-    probe = socket.socket()
-    probe.settimeout(timeout_seconds)
-    try:
-        probe.connect((host, int(port or 22)))
-    finally:
-        probe.close()
-
-
-def _resolve_ssh_options(host, user=None, port=22):
-    resolved = {
-        "host": host,
-        "user": user,
-        "port": port,
-        "key_filename": None,
-    }
-    config_path = os.path.expanduser("~/.ssh/config")
-    if not os.path.exists(config_path):
-        return resolved
-
-    try:
-        ssh_config = paramiko.SSHConfig()
-        with open(config_path, encoding="utf-8") as handle:
-            ssh_config.parse(handle)
-        entry = ssh_config.lookup(host)
-    except Exception:
-        return resolved
-
-    resolved["host"] = entry.get("hostname") or resolved["host"]
-    resolved["user"] = user or entry.get("user") or resolved["user"]
-
-    entry_port = entry.get("port")
-    if entry_port:
-        try:
-            resolved["port"] = int(entry_port)
-        except (TypeError, ValueError):
-            pass
-
-    identity_files = entry.get("identityfile") or []
-    if identity_files:
-        expanded = [os.path.expanduser(path) for path in identity_files]
-        resolved["key_filename"] = expanded if len(expanded) > 1 else expanded[0]
-
-    return resolved
+from tpweb.services.slurm_messages import classify_slurm_resource_message
 
 
 def _record_remote_job(run_id_raw, *, job_id, remote_job_dir):
@@ -256,15 +189,6 @@ def _build_colabfold_config(cfg_dict):
         ).strip(),
         cpu_retry_attempts=_env_int("TPW_COLABFOLD_CPU_RETRIES", default=2),
     )
-
-
-def _config_text(cfg_dict, section, option, default=None):
-    try:
-        value = cfg_dict.get(section, option, fallback=None)
-    except Exception:
-        value = None
-    text = str(value or "").strip()
-    return text or default
 
 
 # ---------------------------------------------------------------------------
@@ -685,45 +609,6 @@ def run_remote_colabfold(cfg_dict, folder_path, genome):
         run_id_raw,
         workers,
     )
-
-
-def _open_remote_session(config):
-    ssh = paramiko.SSHClient()
-    ssh.load_system_host_keys()
-    ssh.set_missing_host_key_policy(paramiko.WarningPolicy())
-    ssh.connect(
-        config.ssh_host,
-        port=config.ssh_port,
-        username=config.ssh_user,
-        password=config.ssh_password,
-        timeout=config.ssh_connect_timeout,
-        banner_timeout=config.ssh_connect_timeout,
-        auth_timeout=config.ssh_connect_timeout,
-        allow_agent=True,
-        look_for_keys=True,
-        key_filename=config.ssh_key_filename,
-    )
-    scp_client = SCPClient(ssh.get_transport())
-    sftp = ssh.open_sftp()
-    return ssh, scp_client, sftp
-
-
-def _close_remote_session(ssh, scp_client, sftp):
-    if scp_client is not None:
-        try:
-            scp_client.close()
-        except Exception:
-            pass
-    if sftp is not None:
-        try:
-            sftp.close()
-        except Exception:
-            pass
-    if ssh is not None:
-        try:
-            ssh.close()
-        except Exception:
-            pass
 
 
 def _candidate_pdb_path(folder_path, locus_tag):
