+++ pipeline/interproscan_remote.py
@@ -1,7 +1,6 @@
 import gzip
 import os
 import shlex
-import socket
 import time
 from dataclasses import dataclass
 from datetime import datetime
@@ -9,89 +8,15 @@ from datetime import datetime
 import paramiko
 from scp import SCPClient
 
-from tpweb.services.slurm_messages import classify_slurm_resource_message
-
-
-REMOTE_FAILURE_PREFIXES = (
-    "FAILED",
-    "CANCELLED",
-    "TIMEOUT",
-    "OUT_OF_MEMORY",
-    "NODE_FAIL",
+from slurm_remote_command import (
+    REMOTE_FAILURE_PREFIXES,
+    _assert_ssh_reachable,
+    _config_text,
+    _env_int,
+    _env_text,
+    _resolve_ssh_options,
 )
-
-
-def _env_int(name, default):
-    raw = os.getenv(name)
-    if raw is None:
-        return default
-    try:
-        return int(raw.strip())
-    except (TypeError, ValueError):
-        return default
-
-
-def _env_text(name, default=None):
-    raw = os.getenv(name)
-    if raw is None:
-        return default
-    text = str(raw).strip()
-    return text or default
-
-
-def _config_text(cfg_dict, section, option, default=None):
-    try:
-        value = cfg_dict.get(section, option, fallback=None)
-    except Exception:
-        value = None
-    text = str(value or "").strip()
-    return text or default
-
-
-def _assert_ssh_reachable(host, port, timeout_seconds):
-    probe = socket.socket()
-    probe.settimeout(timeout_seconds)
-    try:
-        probe.connect((host, int(port or 22)))
-    finally:
-        probe.close()
-
-
-def _resolve_ssh_options(host, user=None, port=22):
-    resolved = {
-        "host": host,
-        "user": user,
-        "port": port,
-        "key_filename": None,
-    }
-    config_path = os.path.expanduser("~/.ssh/config")
-    if not os.path.exists(config_path):
-        return resolved
-
-    try:
-        ssh_config = paramiko.SSHConfig()
-        with open(config_path, encoding="utf-8") as handle:
-            ssh_config.parse(handle)
-        entry = ssh_config.lookup(host)
-    except Exception:
-        return resolved
-
-    resolved["host"] = entry.get("hostname") or resolved["host"]
-    resolved["user"] = user or entry.get("user") or resolved["user"]
-
-    entry_port = entry.get("port")
-    if entry_port:
-        try:
-            resolved["port"] = int(entry_port)
-        except (TypeError, ValueError):
-            pass
-
-    identity_files = entry.get("identityfile") or []
-    if identity_files:
-        expanded = [os.path.expanduser(path) for path in identity_files]
-        resolved["key_filename"] = expanded if len(expanded) > 1 else expanded[0]
-
-    return resolved
+from tpweb.services.slurm_messages import classify_slurm_resource_message
 
 
 def _record_remote_job(run_id_raw, *, job_id, remote_job_dir):
